fix: corriger bugs Vigenère, Permutation, InvalidPathException + ajouter 310 tests PHPUnit
- Vigenère: ord('A') → ord('a') pour position/decalage (cleanText retourne du lowercase), roundtrip fonctionne
- Permutation: (int) ceil/floor pour éviter warnings "String offset cast" en PHP 8.4
- InvalidPathException: type nullable explicite ?Exception, passer $previous au parent
- Tests Phase 1 (Services): CaptchaService, César, Vigenère, Permutation, Hamming
- Tests Phase 2 (Models): File, Bash, InterfaceMail, Instagram, LessonProgress, ResetProgress
- Tests Phase 3 (Controllers): AbstractController, Captcha, RouteAttribute, LeconComplete
- Tests Phase 4 (Router): résolution dynamique, kebabToPascal, params, 404, edge cases
- phpunit.xml: exclure scripts standalone qui interfèrent avec PHPUnit
- bootstrap.php: stub url() pour les controllers

+++ src/Services/Code/Permutation.php
@@ -29,13 +29,13 @@ class Permutation extends AbstractCode
         $text = str_replace(" ", $space_char, $text);
         $text = self::cleanText($text);
 
-        $n_rows = ceil(strlen($text) / $n_columns);
+        $n_rows = (int) ceil(strlen($text) / $n_columns);
         $text .= str_repeat(\mb_strtolower($space_char), $n_columns * $n_rows - strlen($text));
         $order = self::getOrderFromKey($key);
 
         $code = "";
         for ($i = 0; $i < strlen($text); ++$i) {
-            $current_pos = ($i % $n_rows) * $n_columns + $order[floor($i/$n_rows)];   // Formule pour lire de haut en bas plutôt que de gauche à droite
+            $current_pos = ($i % $n_rows) * $n_columns + $order[(int) floor($i/$n_rows)];   // Formule pour lire de haut en bas plutôt que de gauche à droite
             $code .= $text[$current_pos];
         }
 
@@ -59,13 +59,13 @@ class Permutation extends AbstractCode
         $text = strtolower($text);
         $n_columns = strlen($key);
 
-        $n_rows = ceil(strlen($text) / $n_columns);
+        $n_rows = (int) ceil(strlen($text) / $n_columns);
         $order = self::getOrderFromKey($key);
         $order = array_flip($order);
 
         $code = "";
         for ($i = 0; $i < strlen($text); ++$i) {
-            $current_pos = $order[$i%$n_columns] * $n_rows + floor($i/$n_columns);   // Formule pour lire de haut en bas plutôt que de gauche à droite
+            $current_pos = $order[$i % $n_columns] * $n_rows + (int) floor($i / $n_columns);   // Formule pour lire de haut en bas plutôt que de gauche à droite
             $code .= $text[$current_pos];
         }
